Move to async loading of save data
Save data is now loaded asynchronously due to the sheer size of it.
The new SaveDataCallback can be implemented to handle the various stages
of loading the save data.
The default loader (DDSaveLoader) has a basic implementation of this
new callback and the existing CompressionProgressCallback that doesn't
do anything. This can be overriden if needed.

DDSave now populates the inventory data for the characters too.